% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{get5p}
\alias{get5p}
\title{Get a specific S5p product}
\usage{
get5p(month, year, number, id, fn)
}
\arguments{
\item{month}{Desired month in mm format (April would be "04", December would be "12")}

\item{year}{Desired year in yyyy format ("2020" not "20")}

\item{number}{Desired .csv file index (from 1 to the maximum number of files returned by listlast() function)}

\item{id}{Desired S5P file index (from 1 to the maximum number of files returned by get5plist() function)}

\item{fn}{Desired name of the file to be used when saving the product to the local working directory}
}
\description{
The functions works after get5plist() function, it retrives a specific S5P product and allows to save it
}
\examples{
get5p("04","2020","3","1","lastday.ncf")
}
\keyword{list}
