% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{get5p_latlon}
\alias{get5p_latlon}
\title{Get the latest S5p products (lat long search)}
\usage{
get5p_latlon(lat, lon, id = NULL)
}
\arguments{
\item{lat}{Latitude (degrees) to search at}

\item{lon}{Longitude (degrees) to search at}

\item{id=NULL}{If the id parameter is omitted, the function returns a list of up to 10 S5p products available at the given coordinates,
if an id is specified the functions downloads the product with the indicated id. If id=-1 then all the found products will be downloaded automatically.}
}
\description{
The function works by searching the latest 10 products available for a specific location (lat long).
It allows to download a specific product after identifying it.
}
\examples{
get5p_latlon("44","12")
get5p_latlon("44","12",1)
}
\keyword{download}
\keyword{list,}
