% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{get5plist}
\alias{get5plist}
\title{Get a list of 5p products}
\usage{
get5plist(month, year, number)
}
\arguments{
\item{month}{Desired month in mm format (April would be "04", December would be "12")}

\item{year}{Desired year in yyyy format ("2020" not "20")}

\item{number}{Desired .csv file index (from 1 to the maximum number of files returned by listlast() function)}
}
\description{
The functions works after listlast(month, year) function. After retrieving the list of available .csv files with listlast(),
the get5plist() functions allows to retrieve a list of the available products described inside a specific .csv file.
}
\examples{
get5plist("04","2020","3")
}
\keyword{list}
