% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{listlast}
\alias{listlast}
\title{List the latest files available in the Coperniucs 5p catalogue}
\usage{
listlast(month, year)
}
\arguments{
\item{month}{Desired month in mm format (April would be "04", December would be "12")}

\item{year}{Desired year in yyyy format ("2020" not "20")}
}
\description{
This function searches inside the current catalogue of archived products for the Platform Sentinel 5P (https://scihub.copernicus.eu/catalogueview/S5P/)
and returns a list of the archived .csv files of a specific month and year.
}
\examples{
listlast("04","2020")
}
\keyword{list}
