% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{plot5p}
\alias{plot5p}
\title{Get a S5p product}
\usage{
plot5p(fil, variable = NULL)
}
\arguments{
\item{fil}{Name of the file to inspect/plot}

\item{variable}{If the field is NULL then the function will return a list containing all the informations of the .nc file.
Normally the variables names are included as PRODUCT/variable. If the variable name is given as parameter then the function
will plot that variable and save the resulting plot as export.svg. Plotting the variable may require some time, depending on
the size of the downloaded file.}
}
\description{
This function allows to inspect a downloaded S5p file or plots it
}
\examples{
plot5p("S5P_OFFL_L2__AER_LH_20200527T115918_20200527T134048_13579_01_010302_20200529T045906.nc")
plot5p("S5P_OFFL_L2__AER_LH_20200527T115918_20200527T134048_13579_01_010302_20200529T045906.nc","aerosol_index_340_380")
}
\keyword{download,}
\keyword{list,}
\keyword{plot}
